requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
torch
huggingface_hub
langchain
langchain-community
sentence-transformers
chromadb
transformers
PyMuPDF
redis
requests
httpx